added memcache support if memcache url is specified in config file, Shoal will use memcache to store the list of Squids. If no url is specified Shoal will continue to use a list stored in memory that will be specific to the current process. The list stored in memory has not been tested with apache, but for highest performance it is recommended to use memcache.
updated how squids are tracked. removed Squid class as its easier to
keep track of Squid data if it maintains the same format (ie. in a
dictionary) and not have to convert between an object and dictionary
throughout the code.

Added setup.py to include memcache library.
Added new script in scripts/ that should be used if using memcache and
apache. The monitor.py script should be a single process running in the
background, and the shoal_wsgi.py script can be run through Apache
mod_wsgi with multiple processes.